+++ invoicer_project/static/js/utils_clients.js
@@ -18,49 +18,6 @@ export const returnAllFields = function () {
     ];
 }
 
-export function setErrorAttributesToFields(errorsObject) {
-    let fieldIndex = 0;
-    let fields = returnAllFields();
-    for (let error in errorsObject) {
-        if (errorsObject[error]) {
-            fields[fieldIndex].setAttribute("error", "true");
-            fields[fieldIndex].setAttribute("errorText", errorsObject[error]);
-        }
-        fieldIndex++;
-    }
-}
-
-export function clearErrorAttributes(returnAllFieldsList) {
-        for (let field of returnAllFieldsList) {
-            field.addEventListener('input', () => {
-                field.removeAttribute("error");
-                field.removeAttribute("errorText");
-            })
-        }
-}
-
-export function setMaxFieldContainerHeights(returnAllFieldsList) {
-    for (let field of returnAllFieldsList) {
-        field.shadowRoot.querySelector('.md3-text-field__field').shadowRoot.querySelector('.md3-field').querySelector('.md3-field__container').style.maxHeight = "56px";
-    }
-}
-
-export function removeAllErrorAttributes(returnAllFieldsList) {
-    for (let item of returnAllFieldsList) {
-        item.removeAttribute("error");
-        item.removeAttribute("errorText");
-    }
-}
-
-export function allAreFalse(object) {
-    for (let key in object) {
-        if (object[key]) {
-            return false;
-        }
-    }
-    return true;
-}
-
 export function hideUnnecessaryElementsInMenu() {
   const childElements = document.querySelectorAll('.additional_navigation_elements > *');
     childElements.forEach(child => {
